Update drive to use new system
