+++ src/main/java/frc/robot/commands/GoToPosePID.java
@@ -2,7 +2,7 @@ package frc.robot.commands;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.DrivetrainImpl;
 
 import java.util.function.Supplier;
 
@@ -11,14 +11,14 @@ import java.util.function.Supplier;
  */
 public class GoToPosePID extends CommandBase {
 
-    private final Drivetrain drive;
+    private final DrivetrainImpl drive;
 
     private final Supplier<Pose2d> pose;
 
     /**
      * Runs the chassis PIDs to move the robot to a specific pose.
      */
-    public GoToPosePID(Supplier<Pose2d> pose, Drivetrain drive) {
+    public GoToPosePID(Supplier<Pose2d> pose, DrivetrainImpl drive) {
         this.drive = drive;
         this.pose = pose;
