+++ src/main/java/frc/robot/commands/test/GoToPoseTest.java
@@ -6,20 +6,20 @@ import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.constants.miscConstants.TestConstants;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.DrivetrainImpl;
 
 /**
  * Tests the odometry of the robot by driving a certain distance and calculating the error.
  */
 public class GoToPoseTest extends CommandBase {
 
-    private final Drivetrain drive;
+    private final DrivetrainImpl drive;
 
     private double startTime;
     private Pose2d finalPose;
     private Pose2d error;
 
-    public GoToPoseTest(Drivetrain drive) {
+    public GoToPoseTest(DrivetrainImpl drive) {
         this.drive = drive;
 
         addRequirements(drive);
