+++ src/main/java/frc/robot/commands/test/PoseTransform.java
@@ -6,21 +6,21 @@ import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.constants.miscConstants.TestConstants;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.DrivetrainImpl;
 
 /**
  * Tests the odometry of the robot by driving a certain distance and calculating the error.
  */
 public class PoseTransform extends CommandBase {
 
-    private final Drivetrain drive;
+    private final DrivetrainImpl drive;
 
     private double startTime;
     private Pose2d finalPose;
     private final Transform2d distanceToMove;
     private Pose2d error;
 
-    public PoseTransform(Drivetrain drive, Transform2d poseTransform) {
+    public PoseTransform(DrivetrainImpl drive, Transform2d poseTransform) {
         this.drive = drive;
         // finalPose is position after robot moves from current position-- startPose-- by the values that are inputted-- distanceToMove
         distanceToMove = poseTransform;
