+++ src/main/java/frc/robot/commands/test/TestDriveVelocity.java
@@ -6,7 +6,7 @@ import edu.wpi.first.math.util.Units;
 import edu.wpi.first.networktables.GenericEntry;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.constants.miscConstants.TestConstants;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.DrivetrainImpl;
 import frc.robot.subsystems.Module;
 import frc.robot.util.TimeAccuracyTest;
 
@@ -15,11 +15,11 @@ import frc.robot.util.TimeAccuracyTest;
  */
 public class TestDriveVelocity extends CommandBase {
 
-    private final Drivetrain drive;
+    private final DrivetrainImpl drive;
     private final GenericEntry testEntry;
     private final TimeAccuracyTest[] timeAccuracyTests = new TimeAccuracyTest[4];
 
-    public TestDriveVelocity(Drivetrain drive, GenericEntry testEntry) {
+    public TestDriveVelocity(DrivetrainImpl drive, GenericEntry testEntry) {
         this.drive = drive;
         this.testEntry = testEntry;
         addRequirements(drive);
