+++ src/main/java/frc/robot/commands/test/TestSteerAngle.java
@@ -5,7 +5,7 @@ import edu.wpi.first.math.kinematics.SwerveModuleState;
 import edu.wpi.first.networktables.GenericEntry;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.constants.miscConstants.TestConstants;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.DrivetrainImpl;
 import frc.robot.subsystems.Module;
 import frc.robot.util.TimeAccuracyTest;
 
@@ -14,11 +14,11 @@ import frc.robot.util.TimeAccuracyTest;
  */
 public class TestSteerAngle extends CommandBase {
 
-    private final Drivetrain drive;
+    private final DrivetrainImpl drive;
     private final GenericEntry testEntry;
     private final TimeAccuracyTest[] timeAccuracyTests = new TimeAccuracyTest[4];
 
-    public TestSteerAngle(Drivetrain drive, GenericEntry testEntry) {
+    public TestSteerAngle(DrivetrainImpl drive, GenericEntry testEntry) {
         this.drive = drive;
         this.testEntry = testEntry;
         addRequirements(drive);
