+++ src/main/java/frc/robot/commands/vision/TestVisionAlignment.java
@@ -3,7 +3,7 @@ package frc.robot.commands.vision;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.DrivetrainImpl;
 import frc.robot.util.Vision;
 
 /**
@@ -11,7 +11,7 @@ import frc.robot.util.Vision;
  * It only works if it can see an april tag at the setpoint
  */
 public class TestVisionAlignment extends CommandBase {
-    private final Drivetrain drive;
+    private final DrivetrainImpl drive;
     private final Vision vision;
     private final double setpoint;
     private double mostRecentAngle;
@@ -23,7 +23,7 @@ public class TestVisionAlignment extends CommandBase {
      * @param drive       The drivetrain
      * @param vision      The vision
      */
-    public TestVisionAlignment(double targetAngle, Drivetrain drive, Vision vision) {
+    public TestVisionAlignment(double targetAngle, DrivetrainImpl drive, Vision vision) {
         addRequirements(drive);
         setpoint = targetAngle;
         this.drive = drive;
